Add Auction feature and enhance User Dashboard
A new Auction feature was implemented, allowing users to view and participate in auctions. Guest and User dashboard layouts were also enhanced to improve user navigation and access to features. Volt was installed to enhance the implementation of new Livewire views. The LotPolicy was updated to improve security with edit, update, and delete actions.