Add user watchlist feature and enhance datetime inputs
This update introduces a new user watchlist page which displays the list of watched items with an option to remove items from the list. It also corrects the input fields for start and end times in the auction item creation by changing them from 'time' to 'datetime-local'. In addition, it updates filters in AuctionItems to use Carbon for datetime parsing. The Watchlist link in the auction layout has also been updated accordingly.